Removed old github api functions and tests. Created utility.py file and moved utility functions from main.py to utility.py. Updated hardcoded env variables to dynamic variables.
