--- .env.example
@@ -1,7 +1,0 @@
-# GitHub
-GITHUB_PAT=your_github_pat
-GITHUB_OWNER=your_github_username
-GITHUB_REPO=name_of_your_github_repo
-
-# OpenAI
-OPENAI_API_KEY=your_openai_api_key
